refactor(perps): move live price subscription to PerpsRelatedMarkets parent
Lift the usePerpsLivePrices hook from the per-pill wrapper into the
parent component so a single subscription covers all related market
symbols. Remove PerpsPillItemWithLiveData — the parent now merges live
percentChange24h into feedItems and passes enriched data to PerpsPillItem.
