+++ app/components/UI/Perps/components/PerpsPillItem/index.ts
@@ -1,2 +1,1 @@
 export { default as PerpsPillItem } from './PerpsPillItem';
-export { default as PerpsPillItemWithLiveData } from './PerpsPillItemWithLiveData';
